(intermediate-storage-instance-config-2) Substanially restructure ConfigurableClass as ConfigPlugin
Summary: First this is a relatively big change in one diff. I'm up for
splitting this up but I do believe it is manageable as is.

I was sorting through the ConfigurableClass stuff and was a bit
confused. It seemed overly complex for what it was trying to accomplish,
so I decided to rethink this a little bit. I think the core problem
was the multiple inheritance, which muddled the clarity of this
subsystem. This is a capability which is really a lot simpler
and is in the province of the config type system, rather than where
it used to live, in serdes.

A config plugin boils down to a config schema and a function that
takes a value validated by that schema and produces an object. It's
very similar to input schema and output schema actually, except
that the config is only evaluated lazily rather than upfront.

With that new framing, we can express this more simply as a decorated
function. Instead of inheriting from ConfigurableClass you just implement
a function decorated by @config_plugin.

An implementation looks like this:

```
@config_plugin(SystemNamedDict('PostgresRunStorageConfig', {'postgres_url': Field(String)}))
def postgres_event_log_storage_config_plugin(plugin_config):
    return PostgresEventLogStorage(plugin_config['postgres_url'])
```

To do this I moved the "inst_info" functionality up to the
DagsterInstance, which knows whether or not the objects were
constructed via the plugin system or not. This keeps the
fact there is a plugin system at all out of the core objects,
which is preferable in my view.

I was throughly confused by the arbitary kwargs flowed up and down
this codepath. It was unused, so I eliminated it. I think shuffling
the config values to the __init__ method manually is far more explicit
and clear than relying on kwargs explosion.

This ends up simplifying both the implementation (negative LOC)  and
the user-facing API, and is a nice win all-around.

Test Plan: BK. View instance info in Dagit

Reviewers: max, alangenfeld, #ft

Differential Revision: https://dagster.phacility.com/D1306